requirements.txt
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0
langchain-google-vertexai>=2.0.0
langgraph>=0.2.0
pydantic>=2.0.0
tavily-python>=0.3.0
aiohttp>=3.8.0
python-dotenv>=1.0.0